requirements.txt
python-dotenv
requests==2.28.0
fastapi==0.95.1
uvicorn==0.21.1